Time: 113 ms (5.37%), Space: 81.1 MB (53.62%) - LeetHub

+++ 2064-minimized-maximum-of-products-distributed-to-any-store/2064-minimized-maximum-of-products-distributed-to-any-store.java
@@ -23,7 +23,7 @@ class Solution {
     public boolean isItpossible(int mid, int[] quantities, int totalstores){
         int storesneeded = 0;
         for(int i = 0; i < quantities.length; i++){
-            storesneeded += (int) Math.ceil((double) quantities[i]/mid);
+            storesneeded +=  Math.ceil((double) quantities[i]/mid);
         }
         return (storesneeded <= totalstores);
